Clean dup fixtures
