+++ nuclia_e2e/nuclia_e2e/tests/test_default_models.py
@@ -1,8 +1,6 @@
 from collections.abc import AsyncIterator
 from nuclia import sdk
-from nuclia.data import get_async_auth
 from nuclia.sdk.auth import AsyncNucliaAuth
-from nuclia_e2e.tests.conftest import ZoneConfig
 from nuclia_e2e.tests.utils import as_default_generative_model_for_kb
 from nuclia_e2e.tests.utils import clean_ask_test_tasks
 from nuclia_e2e.tests.utils import create_ask_agent
@@ -19,42 +17,6 @@ TEST_ENV = os.environ.get("TEST_ENV")
 # Global variable to know which tasks were created in this test
 # suite so we can clean them up properly on fixture teardown
 _tasks_to_delete: list[str] = []
-
-
-@pytest.fixture
-async def kb_id(regional_api_config: ZoneConfig) -> str:
-    """
-    Fixture to provide the knowledge base ID for the tests.
-    """
-    assert regional_api_config.global_config is not None
-    return regional_api_config.permanent_kb_id
-
-
-@pytest.fixture
-async def zone(regional_api_config: ZoneConfig) -> str:
-    """
-    Fixture to provide the zone slug for the tests.
-    """
-    return regional_api_config.zone_slug
-
-
-@pytest.fixture
-def auth() -> AsyncNucliaAuth:
-    """
-    Fixture to provide the async Nuclia authentication object.
-    """
-    return get_async_auth()
-
-
-@pytest.fixture(autouse=True)
-async def account_id(regional_api_config: ZoneConfig, auth: AsyncNucliaAuth) -> str:
-    """
-    Fixture to provide the account slug for the tests.
-    """
-    assert regional_api_config.global_config is not None
-    account_slug = regional_api_config.global_config.permanent_account_slug
-    sdk.NucliaAccounts().default(account_slug)
-    return auth.get_account_id(account_slug)
 
 
 @pytest.fixture
